Update archetypes structure
