+++ maven/archetypes/actions/src/main/resources/archetype-resources/src/main/resources/actionPath/sample.js
