+++ maven/archetypes/actions/src/main/resources/archetype-resources/src/test/resources/actionPath/SampleTests.js
