+++ maven/archetypes/mixed/src/main/resources/archetype-resources/actions/src/main/resources/actionPath/sample.js
