+++ maven/archetypes/mixed/src/main/resources/archetype-resources/actions/src/test/resources/actionPath/SampleTests.js
