+++ maven/archetypes/mixed/src/main/resources/archetype-resources/workflows/src/main/resources/Workflow/workflowPath/Install.element_info.xml
