+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/catalog-items/Blueprint_Source__Sum two numbers.json
