+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/catalog-items/forms/Blueprint_Source__Sum two numbers.json
@@ -1,0 +1,10 @@
+{
+	"id": "f9f34efe-db20-4cf0-ad90-cea126219181",
+	"name": "Sum two numbers",
+	"form": "{\"layout\":{\"pages\":[{\"id\":\"page_1\",\"title\":\"General\",\"sections\":[{\"id\":\"section_project\",\"fields\":[{\"id\":\"project\",\"display\":\"dropDown\",\"signpostPosition\":\"right-middle\"}]},{\"id\":\"section_deploymentName\",\"fields\":[{\"id\":\"deploymentName\",\"display\":\"textField\",\"signpostPosition\":\"right-middle\"}]},{\"id\":\"section_0\",\"fields\":[{\"id\":\"numberOne\",\"display\":\"decimalField\",\"signpostPosition\":\"right-middle\"}]},{\"id\":\"section_1\",\"fields\":[{\"id\":\"numberTwo\",\"display\":\"decimalField\",\"signpostPosition\":\"right-middle\"}]}],\"state\":{}}]},\"schema\":{\"project\":{\"label\":\"Project\",\"type\":{\"dataType\":\"string\",\"isMultiple\":false},\"valueList\":{\"id\":\"projects\",\"type\":\"scriptAction\"},\"constraints\":{\"required\":true}},\"numberOne\":{\"label\":\"Number One\",\"type\":{\"dataType\":\"decimal\",\"isMultiple\":false},\"default\":1,\"constraints\":{\"required\":true}},\"numberTwo\":{\"label\":\"Number Two\",\"type\":{\"dataType\":\"decimal\",\"isMultiple\":false},\"default\":2,\"constraints\":{\"required\":true}},\"deploymentName\":{\"label\":\"Deployment Name\",\"type\":{\"dataType\":\"string\",\"isMultiple\":false},\"constraints\":{\"required\":true,\"max-value\":80}}},\"options\":{\"externalValidations\":[]}}",
+	"styles": null,
+	"sourceType": "com.vmw.vro.workflow",
+	"type": "requestForm",
+	"status": "ON",
+	"formFormat": "JSON"
+}
