+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/content-sources/Blueprint_Source.json
@@ -1,0 +1,4 @@
+{
+	"name": "Blueprint_Source",
+	"typeId": "com.vmw.blueprint"
+}
