--- maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/content-sources/source.json
@@ -1,4 +1,0 @@
-{
-  "name": "source",
-  "typeId": "com.vmw.blueprint"
-}
