+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/policy/day2-actions/policy.json
@@ -1,6 +1,6 @@
 {
 	"id": "4a68e2b2-fed5-4583-b8ae-d88dc6ee36e7",
-	"name": "D2A1111",
+	"name": "Template Day 2 Actions Policy",
 	"typeId": "com.vmware.policy.deployment.action",
 	"enforcementType": "SOFT",
 	"description": "21345",
