+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/policy/resource-quota/policy.json
@@ -1,6 +1,6 @@
 {
 	"id": "f7cd2b97-d157-4042-96a3-c9c713729d57",
-	"name": "RQ05",
+	"name": "Template Resource Quota Policy",
 	"typeId": "com.vmware.policy.resource.quota",
 	"enforcementType": "HARD",
 	"description": "",
