+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/resource-actions/resourceAction.json
@@ -1,10 +1,10 @@
 {
-  "id": "Cloud.vSphere.Machine.custom.resourceAction",
-  "name": "resourceAction",
-  "displayName": "resourceAction",
-  "providerName": "vro-workflow",
-  "resourceType": "Cloud.vSphere.Machine",
-  "status": "RELEASED",
-  "runnableItem": {},
-  "formDefinition": {}
+	"id": "Cloud.vSphere.Machine.custom.resourceAction",
+	"name": "resource_action",
+	"displayName": "Resource Action",
+	"providerName": "vro-workflow",
+	"resourceType": "Cloud.vSphere.Machine",
+	"status": "RELEASED",
+	"runnableItem": {},
+	"formDefinition": {}
 }
