+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vra/src/main/resources/subscriptions/subscription.json
@@ -1,11 +1,13 @@
 {
 	"id": "sub_uuid",
-    "type": "RUNNABLE",
-    "name": "subscription",
-    "eventTopicId": "compute.allocation.pre",
-    "runnableType": "extensibility.abx",
-    "runnableId": "runnableId",
-    "constraints": {
-    	"projectId": ["projectId"]
-    }
+	"type": "RUNNABLE",
+	"name": "Template Subscription",
+	"eventTopicId": "compute.allocation.pre",
+	"runnableType": "extensibility.abx",
+	"runnableId": "runnableId",
+	"constraints": {
+		"projectId": [
+			"projectId"
+		]
+	}
 }
