--- maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/actions/sample.test.ts
@@ -1,7 +1,0 @@
-import { SampleClass } from "./sample"
-
-describe("Tests", () => {
-    it("should sum two numbers", () => {
-        expect(new SampleClass().sum(1, 2)).toBe(3)
-    })
-})
