+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/elements/config/SampleConfig2.conf.yaml
