--- maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/elements/config/sample.conf.ts
@@ -1,11 +1,0 @@
-import { Configuration } from "vrotsc-annotations";
-
-@Configuration({
-    name: "Sample Config",
-    path: "MyOrg/MyProject"
-})
-export class SampleConfig {
-    field1: string;
-    field2: number;
-    field3: any;
-}
