+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/elements/resource/sample_3.xml
