+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/elements/resource/sample_4.json
