+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/policies/SamplePolicy.pl.ts
