+++ maven/archetypes/ts-vra-ng/src/main/resources/archetype-resources/vro/src/workflows/SampleWorkflow.wf.ts
