+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/actions/SampleClass.test.ts
@@ -1,0 +1,7 @@
+import { SampleClass } from "./SampleClass"
+
+describe("Tests", () => {
+	it("should sum two numbers", () => {
+		expect(new SampleClass().sum(1, 2)).toBe(3)
+	})
+})
