+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/actions/SampleClass.ts
@@ -1,0 +1,5 @@
+export class SampleClass {
+	public sum(x: number, y: number) {
+		return x + y;
+	}
+}
