--- maven/archetypes/ts/src/main/resources/archetype-resources/src/actions/sample.test.ts
@@ -1,7 +1,0 @@
-import { SampleClass } from "./sample"
-
-describe("Tests", () => {
-    it("should sum two numbers", () => {
-        expect(new SampleClass().sum(1, 2)).toBe(3)
-    })
-})
