--- maven/archetypes/ts/src/main/resources/archetype-resources/src/actions/sample.ts
@@ -1,5 +1,0 @@
-export class SampleClass {
-    public sum(x: number, y: number) {
-        return x + y;
-    }
-}
