+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/elements/config/SampleConfig.conf.ts
@@ -1,0 +1,11 @@
+import { Configuration } from "vrotsc-annotations";
+
+@Configuration({
+	name: "Sample Config",
+	path: "MyOrg/MyProject"
+})
+export class SampleConfig {
+	field1: string;
+	field2: number;
+	field3: any;
+}
