+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/elements/config/SampleConfig2.conf.yaml
