+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/elements/resource/sample_2.json.element_info.json
