+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/elements/resource/sample_3.xml.element_info.yaml
