+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/elements/resource/sample_4.json
