+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/policies/SamplePolicy.pl.ts
