+++ maven/archetypes/ts/src/main/resources/archetype-resources/src/workflows/SampleWorkflow.wf.form.json
