+++ maven/archetypes/vrealize/src/main/resources/archetype-resources/vro/actions/src/main/resources/actionPath/sample.js
