+++ maven/archetypes/xml/src/main/resources/archetype-resources/src/main/resources/Workflow/workflowPath/Install.element_info.xml
